requirements.txt
fastapi
uvicorn
pymongo
PyPDF2
requests
python-dotenv
google-api-python-client
